pyproject.toml
[tool.poetry]
name = "talon-of-the-hawk"
version = "0.1.0"
description = "My custom talon scripts"
authors = ["Sebastian Battle <sabattle@eckerd.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.17.0"
flake8 = "^4.0.1"
black = "^21.12b0"
commitizen = "^2.20.4"
mypy = "^0.931"
types-PyYAML = "^6.0.3"

[tool.black]
line-length = 120

[tool.mypy]
show_error_codes = true
ignore_missing_imports = true
disallow_untyped_defs = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
